Created user create and update DTOs, updated the user service, ready for review later

+++ TravelBuddyApi/Services/Implementations/UserService.cs
@@ -1,3 +1,4 @@
+using TravelBuddyApi.DTOs;
 using TravelBuddyApi.Models;
 using TravelBuddyApi.Repositories.Abstract;
 using TravelBuddyApi.Services.Interfaces;
@@ -11,18 +12,55 @@ public class UserService(IUserRepository _userRepository)
         return await _userRepository.GetUserByIdAsync(userId);
     }
 
-    public async Task AddUserAsync(User user)
+    public async Task AddUserAsync(long userId, UserCreateDTO userCreateDTO)
     {
-        await _userRepository.AddUserAsync(user);
+        var newUser = new User
+        {
+            UserId = userId,
+            FirstName = userCreateDTO.FirstName,
+            LastName = userCreateDTO.LastName,
+            UserName = userCreateDTO.UserName,
+            DateOfBirth = userCreateDTO.DateOfBirth,
+            Gender = userCreateDTO.Gender,
+            Nationality = userCreateDTO.Nationality,
+            EmailAddress = userCreateDTO.EmailAddress,
+            PasswordHash = userCreateDTO.PasswordHash,
+            ProfileInfo = userCreateDTO.ProfileInfo,
+            ProfileImageUrl = userCreateDTO.ProfileImageUrl
+        };
+
+        await _userRepository.AddUserAsync(newUser);
     }
 
-    public async Task UpdateUserAsync(User user)
+    public async Task UpdateUserAsync(long userId, UserUpdateDTO userUpdateDTO)
     {
-        await _userRepository.UpdateUserAsync(user);
+        User getUser = await _userRepository.GetUserByIdAsync(userId);
+        if (getUser == null)
+        {
+            throw new InvalidOperationException("The requested user does not exist");
+        }
+
+        getUser.FirstName = userUpdateDTO.FirstName;
+        getUser.LastName = userUpdateDTO.LastName;
+        getUser.UserName = userUpdateDTO.UserName;
+        getUser.DateOfBirth = userUpdateDTO.DateOfBirth;
+        getUser.Gender = userUpdateDTO.Gender;
+        getUser.Nationality = userUpdateDTO.Nationality;
+        getUser.EmailAddress = userUpdateDTO.EmailAddress;
+        getUser.PasswordHash = userUpdateDTO.PasswordHash;
+        getUser.ProfileInfo = userUpdateDTO.ProfileImageUrl;
+        getUser.ProfileImageUrl = userUpdateDTO.ProfileImageUrl;
+
+        await _userRepository.UpdateUserAsync(getUser);
     }
 
     public async Task RemoveUserAsync(long userId)
     {
+        User getUser = await _userRepository.GetUserByIdAsync(userId);
+        if (getUser == null)
+        {
+            throw new InvalidOperationException("The requested user does not exist");
+        }
         await _userRepository.RemoveUserAsync(userId);
     }
 }
